# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Ticketing System API",
        "description": "API documentation for Ticketing System",
        "contact": {
            "email": "support@ticketingsystem.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/vip-management/vips": {
            "post": {
                "tags": [
                    "VIP Management"
                ],
                "summary": "Create a new VIP member",
                "description": "Store a newly created VIP member in storage.",
                "operationId": "a8ad9165353e8b51a1feb89b3510be70",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "card_number",
                                    "name",
                                    "validity_start",
                                    "validity_end"
                                ],
                                "properties": {
                                    "card_number": {
                                        "description": "VIP card number (max 10 digits)",
                                        "type": "string",
                                        "example": "1234567890"
                                    },
                                    "name": {
                                        "description": "VIP member name",
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "validity_start": {
                                        "description": "Validity start date",
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-01-01"
                                    },
                                    "validity_end": {
                                        "description": "Validity end date",
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-12-31"
                                    },
                                    "status": {
                                        "description": "VIP status",
                                        "type": "string",
                                        "example": "Active"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "VIP member created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/vip-management/vips/{id}": {
            "put": {
                "tags": [
                    "VIP Management"
                ],
                "summary": "Update an existing VIP member",
                "description": "Update the specified VIP member in storage.",
                "operationId": "b27152c4be5b4f13c391be7c5a85a594",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "VIP member ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "card_number",
                                    "name",
                                    "validity_start",
                                    "validity_end"
                                ],
                                "properties": {
                                    "card_number": {
                                        "description": "VIP card number (max 10 digits)",
                                        "type": "string",
                                        "example": "1234567890"
                                    },
                                    "name": {
                                        "description": "VIP member name",
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "validity_start": {
                                        "description": "Validity start date",
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-01-01"
                                    },
                                    "validity_end": {
                                        "description": "Validity end date",
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-12-31"
                                    },
                                    "status": {
                                        "description": "VIP status",
                                        "type": "string",
                                        "example": "Active"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "VIP member updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "VIP member not found"
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/vip-management/vips/expiring": {
            "get": {
                "tags": [
                    "VIP Management"
                ],
                "summary": "Get VIP members expiring soon",
                "description": "Get expiring VIP members.",
                "operationId": "221406d35a5616b7d27bc8f8aa0f88e3",
                "responses": {
                    "200": {
                        "description": "List of expiring VIP members",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}